Auto create user on sign in fail + email submission MVP working
This commit solves two issues:
1. If using the stored cookie to sign in fails, the site automatically
creates a new user rather than just failing. This is much more graceful.
2. When a classifier is submitted, a prompt to enter an email where the
classifier to be sent to is shown to the user.

When the user clicks Submit, here’s what happens:
1. Check for whether the User has an email (for now, I have it set so
that you always enter an email because there is currently no way to
update an email on a user elsewhere if you make a typo or something)
2. Prompt the user to enter an email
3. Validate email, show error banner if validation fails
4. If validation passes, make a PUT request to update the User’s email
and then proceed with submitting the classifier